+++ NEWS.md
@@ -8,6 +8,7 @@
   before `pos_*` was excluded for only one of them
 * random effects: it is now possible to use different grouping levels in
   different sub-models (when providing a list of model formulas)
+* `predDF()` bug fix: the parameter for all methods is now called `object`
 
 --------------------------------------------------------------------------------